# specs/bff-openapi.yaml
openapi: 3.1.0
info:
  title: Fantasy Hockey BFF API
  description: Backend for Frontend aggregation layer for the Fantasy Hockey application
  version: "1.0"
servers:
- url: /
security:
- bearerAuth: []
...
paths:
  /api/v1/projections/{id}:
    get:
      tags:
      - Projections
      summary: "Fetch one of the current user's saved projections, including its data"
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Projection found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "404":
          description: No such projection for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
    put:
      tags:
      - Projections
      summary: Update one of the current user's saved projections
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectionRequest"
        required: true
      responses:
        "200":
          description: Projection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "404":
          description: No such projection for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "409":
          description: Another projection with that name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
    delete:
      tags:
      - Projections
      summary: Delete one of the current user's saved projections
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Projection deleted
        "404":
          description: No such projection for this user
...
  /api/v1/projections:
    get:
      tags:
      - Projections
      summary: List the current user's saved projections (metadata only)
      operationId: list
      responses:
        "200":
          description: "The user's projections, newest first"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectionSummaryResponse"
    post:
      tags:
      - Projections
      summary: Save a new projection for the current user
      description: "Send `source` with an empty `data.players` to have the server\
        \ fill the player rows in from its own read model, instead of uploading ~1600\
        \ players the client just downloaded. Send `data.players` (and no `source`)\
        \ when the rows are user-specific, e.g. copied or carried over from the demo."
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectionRequest"
        required: true
      responses:
        "201":
          description: Projection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
        "409":
          description: A projection of that kind already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectionResponse"
...
  /api/v1/projection-model/splits/skaters:
    get:
      tags:
      - Projection model
      summary: Skaters' measured totals over a range of a season's games
      description: "What players actually produced over a stretch of their team's\
        \ schedule — measured, not projected. Ranges are in team game numbers, so\
        \ the same range covers the same stretch for every player; one who missed\
        \ some of them shows fewer games. Give either fromGame/toGame or lastGames;\
        \ with neither, the whole season is covered."
      operationId: skaterSplits
      parameters:
      - name: season
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: fromGame
        in: query
        description: First team game in the range (1-based)
        required: false
        schema:
          type: integer
          format: int32
          maximum: 84
          minimum: 1
      - name: toGame
        in: query
        description: "Last team game in the range, inclusive"
        required: false
        schema:
          type: integer
          format: int32
          maximum: 84
          minimum: 1
      - name: lastGames
        in: query
        description: Shorthand for the team's final N games
        required: false
        schema:
          type: integer
          format: int32
          maximum: 84
          minimum: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 100
          maximum: 1000
          minimum: 1
      responses:
        "200":
          description: "Totals over the range, highest scoring first"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerSplitResponse"
  /api/v1/projection-model/splits/goalies:
    get:
      tags:
      - Projection model
      summary: Goalies' measured totals over a range of a season's games
      operationId: goalieSplits
      parameters:
      - name: season
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: fromGame
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 84
          minimum: 1
      - name: toGame
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 84
          minimum: 1
      - name: lastGames
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 84
          minimum: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 100
          maximum: 1000
          minimum: 1
      responses:
        "200":
          description: Totals over the range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerSplitResponse"
  /api/v1/projection-model/seed:
    get:
      tags:
      - Projection model
      summary: Generate projection lines for every mapped player
      description: "Returns the model's projected stat lines keyed by this platform's\
        \ player id, ready to be saved as a new projection. Scoring settings are not\
        \ included — those belong to the user's league, so the client supplies them\
        \ when saving."
      operationId: seed
      parameters:
      - name: season
        in: query
        description: Season to project; defaults to the configured one
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Generated lines plus a summary of what was skipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeededProjectionResponse"
...
components:
  schemas:
    DraftPick:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
        teamId:
          type: string
          maxLength: 64
          minLength: 0
      required:
      - playerId
      - teamId
    DraftState:
      type: object
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/DraftTeam"
          maxItems: 32
          minItems: 0
        order:
          type: array
          items:
            type: string
            maxLength: 64
            minLength: 0
          maxItems: 32
          minItems: 0
        picks:
          type: array
          items:
            $ref: "#/components/schemas/DraftPick"
        finishedAt:
          type: string
          format: date-time
      required:
      - order
      - picks
      - teams
    DraftTeam:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
          minLength: 0
        name:
          type: string
          maxLength: 64
          minLength: 0
        mine:
          type: boolean
      required:
      - id
      - mine
      - name
    PlayerProjection:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
        type:
          type: string
          enum:
          - skater
          - goalie
        stats:
          $ref: "#/components/schemas/PlayerStats"
      required:
      - playerId
      - stats
      - type
    PlayerStats:
      type: object
      properties:
        utility:
          type: object
          additionalProperties:
            type: number
            format: double
        scoring:
          type: object
          additionalProperties:
            type: number
            format: double
      required:
      - scoring
      - utility
    ProjectionSettings:
      type: object
      properties:
        scoringType:
          type: string
          enum:
          - points
          - category
        statWeights:
          type: object
          additionalProperties:
            type: number
            format: double
        activeScoringColumns:
          type: array
          items:
            type: string
        activeUtilityColumns:
          type: array
          items:
            type: string
        scaleSettings:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ScaleConfig"
        decimalSettings:
          type: object
          additionalProperties:
            type: integer
            format: int32
        useDefaultDecimals:
          type: boolean
        leagueSize:
          type: integer
          format: int32
          maximum: 30
          minimum: 2
        rosterSlots:
          $ref: "#/components/schemas/RosterSlots"
        minGoalieGames:
          type: integer
          format: int32
          maximum: 82
          minimum: 0
        yahooSync:
          $ref: "#/components/schemas/YahooSync"
      required:
      - activeScoringColumns
      - activeUtilityColumns
      - decimalSettings
      - scaleSettings
      - scoringType
      - statWeights
      - useDefaultDecimals
    RosterSlots:
      type: object
      properties:
        c:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        lw:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        rw:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        d:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        util:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        bn:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
        g:
          type: integer
          format: int32
          maximum: 50
          minimum: 0
      required:
      - bn
      - c
      - d
      - g
      - lw
      - rw
      - util
    ScaleConfig:
      type: object
      properties:
        scale:
          type: boolean
        scalableStats:
          type: array
          items:
            type: string
      required:
      - scalableStats
      - scale
    UpdateProjectionData:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/ProjectionSettings"
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerProjection"
        draft:
          $ref: "#/components/schemas/DraftState"
        tings:
          $ref: "#/components/schemas/UpdateProjectionData"
          writeOnly: true
      required:
      - settings
    UpdateProjectionRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        data:
          $ref: "#/components/schemas/UpdateProjectionData"
      required:
      - data
      - name
    YahooSync:
      type: object
      properties:
        leagueName:
          type: string
          maxLength: 200
          minLength: 0
        leagueKey:
          type: string
          maxLength: 100
          minLength: 0
        syncedAt:
          type: string
          format: date-time
      required:
      - leagueKey
      - leagueName
      - syncedAt
    ProjectionData:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/ProjectionSettings"
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerProjection"
        draft:
          $ref: "#/components/schemas/DraftState"
        tings:
          $ref: "#/components/schemas/ProjectionData"
          writeOnly: true
      required:
      - players
      - settings
    ProjectionResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum:
          - projection
          - preset_draft
        season:
          type: string
          enum:
          - "20252026"
          - "20262027"
          - "20272028"
          - "20282029"
        data:
          $ref: "#/components/schemas/ProjectionData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - data
      - id
      - kind
      - name
      - season
      - updatedAt
...
    CreateProjectionRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        kind:
          type: string
          description: What the projection is for. Defaults to the user's own.
          enum:
          - projection
          - preset_draft
        data:
          $ref: "#/components/schemas/ProjectionData"
        source:
          type: string
          description: Fill data.players from the server's player read model instead
            of sending them. Requires data.players to be empty; omit it to send your
            own rows.
          enum:
          - default
          - blank
      required:
      - data
      - name
...
    ProjectionSummaryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum:
          - projection
          - preset_draft
        season:
          type: string
          enum:
          - "20252026"
          - "20262027"
          - "20272028"
          - "20282029"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        draftStatus:
          type: string
          enum:
          - none
          - in_progress
          - finished
      required:
      - createdAt
      - draftStatus
      - id
      - kind
      - name
      - season
      - updatedAt
    PlayerSplitResponse:
      type: object
      description: A player's measured totals over a range of their team's games
      properties:
        playerId:
          type: integer
          format: int32
          description: The platform's player id
        name:
          type: string
        teamAbbrev:
          type: string
        type:
          type: string
          description: skater or goalie
        games:
          type: integer
          format: int32
          description: Games the player appeared in within the range
        firstTeamGame:
          type: integer
          format: int32
          description: First team game number covered
        lastTeamGame:
          type: integer
          format: int32
          description: Last team game number covered
        stats:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Totals over the range
      required:
      - games
      - name
      - playerId
      - stats
      - type
    SeededProjectionResponse:
      type: object
      description: "Projection lines generated by the model, keyed by the platform's\
        \ player id"
      properties:
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerProjection"
        season:
          type: integer
          format: int32
          description: Season the lines project
        modelVersion:
          type: string
          description: Model version behind them
        skaters:
          type: integer
          format: int32
        goalies:
          type: integer
          format: int32
        unmapped:
          type: integer
          format: int32
          description: Projected players with no counterpart on this platform. Mostly
            fringe and long-term absent players the platform doesn't carry.
        goaliesWithoutWorkload:
          type: integer
          format: int32
          description: "Goalies the model projects no starts for. Left out rather\
            \ than seeded with a zero save percentage, which would read as the worst\
            \ goalie in the league."
      required:
      - goalies
      - goaliesWithoutWorkload
      - modelVersion
      - players
      - season
      - skaters
      - unmapped
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT